Using the compass to save x and y values of "Nodes" can use these values to point ourselves in the right direction and travel between them
I want to use pythag + trig to constantly find the distance and angle to next node however, the compass gives our massively varying number while the motors are running. So it would require a stop to recalibrate. Might be doable if we do the expected distance then stop and check we actually made it to where we want to be.